Remove border-b when showing settings, refactor logic to prevent fire on gear click, move termInputs into ColumnSettings
