+++ src/components/settings/ColumnSettings.tsx
@@ -1,45 +1,51 @@
-import { useEffect } from "react";
-import { useAppSelector } from "../../app/hooks";
+import { useEffect, useState } from "react";
+import { useAppDispatch, useAppSelector } from "../../app/hooks";
 import {
+  defaultLongTerm,
+  defaultMediumTerm,
   selectTerm,
   selectVisualSettings,
 } from "../../features/settings/settingsSlice";
 import { SetBooleanState } from "./term/WorkDay";
-import { TermInputsEvolved } from "./term/TermInputsEvolved";
+import { saveTerm } from "../../features/settings/utils";
+import { DateTime } from "luxon";
+import Icon from "../atoms/Icon";
+import { UnitType } from "../../features/settings/types";
+import { faArrowRightFromBracket } from "@fortawesome/free-solid-svg-icons/faArrowRightFromBracket";
+import { TermName } from "./term/TermName";
+import RadioButton from "../atoms/RadioButton";
+import { SelectDate } from "./term/SelectDate";
+import { Duration } from "./term/Duration";
+
+const defaultTerms: Record<number, UnitType> = {
+  1: defaultMediumTerm,
+  2: defaultLongTerm,
+};
 
 const handleClickOutside =
   (
     setHideSettings: SetBooleanState,
-    { current: currentSettings }: React.MutableRefObject<HTMLDivElement | null>,
-    { current: currentIcon }: React.MutableRefObject<HTMLElement | null>
+    { current: currentSettings }: React.MutableRefObject<HTMLDivElement | null>
   ) =>
   (event: Event) => {
     if (currentSettings && !currentSettings.contains(event.target as Node)) {
-      if (currentIcon && !currentIcon.contains(event.target as Node)) {
-        setHideSettings(true);
-      }
+      setHideSettings(true);
     }
   };
 
 export const ColumnSettings = ({
   settingsContainer,
-  settingsIcon,
   hideSettings,
   setHideSettings,
   groupId,
 }: {
   settingsContainer: React.MutableRefObject<HTMLDivElement | null>;
-  settingsIcon: React.MutableRefObject<HTMLElement | null>;
   hideSettings: boolean;
   setHideSettings: SetBooleanState;
   groupId: number;
 }) => {
   useEffect(() => {
-    const clickHandler = handleClickOutside(
-      setHideSettings,
-      settingsContainer,
-      settingsIcon
-    );
+    const clickHandler = handleClickOutside(setHideSettings, settingsContainer);
     document.addEventListener("click", clickHandler);
     return () => {
       document.removeEventListener("click", clickHandler);
@@ -48,7 +54,51 @@ export const ColumnSettings = ({
   }, []);
 
   const { bgColor } = useAppSelector(selectVisualSettings);
-  const thisTerm = useAppSelector(selectTerm(groupId));
+  const termData = useAppSelector(selectTerm(groupId));
+
+  const { secondFontColor } = useAppSelector(selectVisualSettings);
+  const dispatch = useAppDispatch();
+  const [title, setTitle] = useState("");
+  const [unitType, setUnitType] = useState("");
+  const [isDuration, setIsDuration] = useState(true);
+  const [startDate, setStartDate] = useState("");
+  const [endDate, setEndDate] = useState("");
+  const [duration, setDuration] = useState(termData.duration);
+  const [repeat, setRepeat] = useState(true);
+
+  useEffect(() => {
+    setUnitType(termData.title.toLowerCase());
+    setTitle(termData.title);
+    setStartDate(termData.startDate ?? "");
+    setEndDate(
+      termData?.endDate ??
+        DateTime.fromISO(termData.startDate ?? "")
+          .plus({ [termData.duration.unit]: termData.duration.qty })
+          .toISO() ??
+        ""
+    );
+    setRepeat(termData.repeat);
+  }, [termData]);
+
+  // const checkboxId = `${category}_repeat-duration`;
+
+  const onChange = (changed: Parameters<typeof saveTerm>) => {
+    if (duration.qty) {
+      saveTerm({
+        enabled: true,
+        isDuration,
+        duration,
+        startDate,
+        dispatch,
+        groupId,
+        unitType,
+        title,
+        repeat,
+        endDate,
+        ...changed,
+      });
+    }
+  };
 
   return (
     <div
@@ -60,19 +110,91 @@ export const ColumnSettings = ({
     >
       <h1 className="font-bold text-2xl leading-none" id="settings-title">
         Customization
+        <Icon
+          onClick={() => {
+            setStartDate(defaultTerms[groupId].startDate ?? "");
+            setEndDate(defaultTerms[groupId].endDate ?? "");
+            setUnitType(defaultTerms[groupId].unitType);
+            setTitle(defaultTerms[groupId].title);
+            setRepeat(defaultTerms[groupId].repeat);
+            setDuration(defaultTerms[groupId].duration);
+          }}
+          data-testid={`${groupId}-restore-defaults`}
+          icon={faArrowRightFromBracket}
+          title="Restore Defaults"
+          faClassName="text-base float-right"
+          faStyle={{ color: secondFontColor }}
+        />
       </h1>
       <hr />
       <div className="my-4 mx-0">
-        <TermInputsEvolved groupId={groupId} termData={thisTerm} />
-      </div>{" "}
-      <div>
-        <button
-          onClick={() => setHideSettings(true)}
-          className="border border-current cursor-pointer pt-0.5 px-1"
-          data-testid="hide-button"
-        >
-          ( Hide )
-        </button>
+        <div className="my-2 mx-0">
+          <div className="my-2 mx-0">
+            <TermName
+              groupId={groupId}
+              title={title}
+              enabled={true}
+              setTitle={setTitle}
+              setUnitType={setUnitType}
+              onChange={onChange}
+            />
+            <div className="inline-block w-1/2">
+              <RadioButton
+                enabled={true}
+                groupId={groupId}
+                firstRadioName="duration"
+                secondRadioName="date"
+                firstIsChecked={isDuration}
+                setIsChecked={setIsDuration}
+              />
+              {/* Temporarily disabled
+          &nbsp; &nbsp;
+          <CheckBox
+            nameId={checkboxId}
+            checked={termData.repeat}
+            onChange={onRepeat(dispatch, category)}
+            disabled={!(enabled && isDuration)}
+            inputStyle={{
+              visibility: "visible",
+              margin: "0px 0.3rem",
+              float: "right",
+            }}
+            labelClass="float-right bottom-[-2px] align-checkbox-label align-middle relative"
+            labelText="Repeat?"
+            labelOnRight
+          /> */}
+            </div>
+          </div>
+          <SelectDate
+            title="Beginning"
+            groupId={groupId}
+            date={startDate}
+            limit={{ max: endDate }}
+            setStartDate={setStartDate}
+            setEndDate={setEndDate}
+            onChange={onChange}
+          />
+          {!isDuration && (
+            <SelectDate
+              title="End"
+              groupId={groupId}
+              date={endDate}
+              limit={{ min: startDate }}
+              setStartDate={setStartDate}
+              setEndDate={setEndDate}
+              onChange={onChange}
+            />
+          )}
+          {isDuration && (
+            <Duration
+              duration={duration}
+              groupId={groupId}
+              enabled={true}
+              setDuration={setDuration}
+              onChange={onChange}
+            />
+          )}
+        </div>
       </div>
     </div>
   );
